Finished Implementation and Major Cleanup
Consider this the first commit with completed code.

Successfully applied IG to segmentation. The user selects a single pixel and for a given/calculated class, the code determines the IG for that pixel and class. The user can select 2 pixels and then for each, choose a class selection method (preset class, nth ranked class, or same as the other pixel class).

There are a lot of messy hacks in the code, such as loading a model checkpoint originally intended for Pytorch-Lightning, the method to draw a red square on the images, and pixel selection.