Bug Fixes
There are two main bug fixes.

The first is that the previous code was never actually calculating integrated gradients. I forgot to update the scaled image at each iteration so it was repeatedly calculating gradients for the original image and then scaling that by original - baseline. Interesting that it actually gave a result...

The second is that I was using cv2 (BGR images) whereas the original SegFormer model uses PIL (RGB images). The swap caused images and gradients to have reversed color channels. I cleaned it up and adopted all PIL, except for one part in Visualization.py where I call this out.